requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
croniter==6.0.0
distro==1.9.0
dnspython==2.8.0
fastapi==0.119.0
google-auth==2.41.1
google-genai==1.45.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.0
openai==2.4.0
pdf2image==1.17.0
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.2
pydantic_core==2.41.4
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
redis==6.4.0
requests==2.32.5
rq==2.6.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
